+++ CHANGELOG.md
@@ -11,6 +11,9 @@ Versions since `1.7.0` only track ABI breaks and not API breaks.
 - Libtcod can now be built without SDL by passing the `-DLIBTCOD_SDL2=disable` flag to CMake.
 - SDL-based accumulate functions can now handle a non-window render target.
 
+### Deprecated
+- Older REXPaint functions using `TCOD_list_t` have been deprecated.
+
 ## [1.20.1] - 2022-01-07
 ### Fixed
 - BDF files with blank lines no longer fail to load with an "Unknown keyword" error.